Move core domain detector classes to new module.
Moving abstract detector types, registry, etc. to new dedicated module.
